requirements.txt
pyTelegramBotAPI
google-genai
firecrawl-py
python-dotenv